more work on firebase activity tracking
