+++ app/src/main/java/com/example/architectureproject/tracking/demo/DummyTrackingDataProvider.kt
@@ -3,8 +3,8 @@ package com.example.architectureproject.tracking.demo
 import com.example.architectureproject.GreenTraceProviders
 import com.example.architectureproject.tracking.TrackingActivity
 import com.example.architectureproject.tracking.TrackingDataGranularity
+import com.example.architectureproject.tracking.TrackingDataHelpers
 import com.example.architectureproject.tracking.TrackingDataProvider
-import com.example.architectureproject.tracking.TrackingDataType
 import com.example.architectureproject.tracking.TrackingEntry
 import com.example.architectureproject.tracking.TrackingPeriod
 import java.time.ZonedDateTime
@@ -50,53 +50,16 @@ class DummyTrackingDataProvider : TrackingDataProvider {
         activitiesByDay[dayOf(new.date)]!![id] = new
     }
 
-    private fun periodOperator(granularity: TrackingDataGranularity) =
-        when (granularity) {
-            TrackingDataGranularity.Day -> TrackingPeriod::dayOf
-            TrackingDataGranularity.Week -> TrackingPeriod::weekOf
-            TrackingDataGranularity.Month -> TrackingPeriod::monthOf
-            TrackingDataGranularity.Year -> TrackingPeriod::yearOf
-        }
-
-    private fun fillGapsOrdered(period: TrackingPeriod,
-                         granularity: TrackingDataGranularity,
-                         map: Map<Long, List<TrackingEntry>>): Map<Long, List<TrackingEntry>> {
-        val periodOp = periodOperator(granularity)
-        var cur = periodOp(period.start).start
-        val periods = linkedMapOf<Long, List<TrackingEntry>>()
-        while (cur < period.end) {
-            periods[cur.toEpochSecond()] = listOf(
-                TrackingEntry(periodOp(cur), TrackingDataType.Emissions, 0f))
-            cur = when (granularity) {
-                TrackingDataGranularity.Day -> cur::plusDays
-                TrackingDataGranularity.Week -> cur::plusWeeks
-                TrackingDataGranularity.Month -> cur::plusMonths
-                TrackingDataGranularity.Year -> cur::plusYears
-            }(1)
-        }
-
-        periods.putAll(map)
-        return periods
-    }
-
     override suspend fun getImpact(
         period: TrackingPeriod,
         granularity: TrackingDataGranularity,
         cid: String
     ): List<TrackingEntry> =
         getActivities(period, cid)
-            .map {
-                GreenTraceProviders.impactProvider.computeImpact(it).let {
-                    it.copy(period = periodOperator(granularity)(it.period.start))
-                }
-            }
-            .groupBy { it.period.start.toEpochSecond() }
-            .let { fillGapsOrdered(period, granularity, it) }
-            .map {
-                it.value[0].copy(value = it.value.fold(0f) { acc, cur ->
-                    acc + cur.value
-                })
-            }
+            .let { TrackingDataHelpers.aggregateImpact(
+                it, GreenTraceProviders.impactProvider::computeImpact,
+                period, granularity
+            ) }
 
     override suspend fun getActivities(period: TrackingPeriod, cid: String) =
         if (cid.isEmpty()) {
